Refactored router to move functions to utils (first step).
